+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/EmptyIndexInput.java
@@ -1,9 +1,9 @@
 /*
- * LuceneOptimizedWrappedIndexInput.java
+ * EmptyIndexInput.java
  *
  * This source file is part of the FoundationDB open source project
  *
- * Copyright 2015-2021 Apple Inc. and the FoundationDB project authors
+ * Copyright 2015-2023 Apple Inc. and the FoundationDB project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,60 +18,54 @@
  * limitations under the License.
  */
 
-package com.apple.foundationdb.record.lucene.codec;
+package com.apple.foundationdb.record.lucene.directory;
 
-import com.google.common.base.Suppliers;
 import org.apache.lucene.store.IndexInput;
-import javax.annotation.Nonnull;
+
+import java.io.EOFException;
 import java.io.IOException;
-import java.util.function.Supplier;
 
 /**
- * An {@code IndexInput} optimized for FDB storage.
+ * An {@link IndexInput} to go with {@link EmptyIndexOutput}.
+ * This input always has length <em>0</em>.
  */
-public class LuceneOptimizedWrappedIndexInput extends IndexInput {
-    Supplier<byte[]> value;
-    private int position;
-
-    public LuceneOptimizedWrappedIndexInput(@Nonnull String name, @Nonnull Supplier<byte[]> supplier) {
-        super(name);
-        value = Suppliers.memoize( () -> supplier.get());
-        position = 0;
+public class EmptyIndexInput extends IndexInput {
+    public EmptyIndexInput(final String resourceDescription) {
+        super(resourceDescription);
     }
 
     @Override
     public void close() throws IOException {
-        // Simply to override the abstract method of super class. Nothing is needed when closing this.
+        // we have no resources to close
     }
 
     @Override
     public long getFilePointer() {
-        return position;
+        return 0;
     }
 
     @Override
     public void seek(final long pos) throws IOException {
-        position = (int) pos;
+        throw new EOFException();
     }
 
     @Override
     public long length() {
-        return value.get().length;
+        return 0;
     }
 
     @Override
     public IndexInput slice(final String sliceDescription, final long offset, final long length) throws IOException {
-        throw new UnsupportedOperationException("Not Supported");
+        throw new EOFException();
     }
 
     @Override
     public byte readByte() throws IOException {
-        return value.get()[position++];
+        throw new EOFException();
     }
 
     @Override
     public void readBytes(final byte[] b, final int offset, final int len) throws IOException {
-        System.arraycopy(value.get(), position, b, offset, len);
-        position = position + len;
+        throw new EOFException();
     }
 }